restructured project to follow Prof's example, fixed imports
