+++ src/procrastination_helper/__init__.py
@@ -1,0 +1,12 @@
+"""
+Procrastination Helper - A fun package to help procrastinators with excuses, quotes, and more.
+
+This package provides various functions to generate procrastination quotes, excuses,
+and other fun tools for when you're not quite ready to tackle that task.
+"""
+from procrastination_helper.quotes import get as quotes
+from procrastination_helper.get_excuse import get_excuse
+from procrastination_helper.customexcuses import get_custom_excuse
+from procrastination_helper.list_of_excuses import get_excuse_list
+
+__version__ = "0.1.0"
